restructure test data structure yet again
