+++ tests/test_calculate.py
@@ -5,9 +5,10 @@ import polars.testing as pl_testing
 import pytest
 
 from athena.utils import calculate
-from tests.test_data import (
-    calculate_min_mean_max_data,
-    calculate_total_percent_coverage_data,
+
+from tests.test_data.calculate import (
+    min_mean_max_data,
+    total_percent_coverage_data,
 )
 
 
@@ -18,15 +19,15 @@ class TestRegionCoverage:
 class TestTotalPctCoverage:
     """
     Data and fixture(s) for the following tests are stored in
-    tests/test_data/calculate_total_percent_coverage_data.py
+    tests/test_data/calculate/total_percent_coverage_data.py
     """
 
     @pytest.mark.parametrize(
         "dataframe, threshold, expected_pct",
         [
-            (calculate_total_percent_coverage_data.dataframe(), 20, 100.00),
-            (calculate_total_percent_coverage_data.dataframe(), 30, 50.00),
-            (calculate_total_percent_coverage_data.dataframe(), 40, 0.00),
+            (total_percent_coverage_data.dataframe(), 20, 100.00),
+            (total_percent_coverage_data.dataframe(), 30, 50.00),
+            (total_percent_coverage_data.dataframe(), 40, 0.00),
         ],
     )
     def test_percentage_correct_against_threshold(
@@ -55,7 +56,7 @@ class TestTotalPctCoverage:
 
     def test_coverage_at_99_99_does_not_round_to_100_pct(self):
         calculated_pct = calculate.total_pct_coverage(
-            coverage_data=calculate_total_percent_coverage_data.dataframe_99_99_pct(),
+            coverage_data=total_percent_coverage_data.dataframe_99_99_pct(),
             threshold=20,
         )
 
@@ -65,7 +66,7 @@ class TestTotalPctCoverage:
 class TestMinMeanMax:
     """
     Data and fixture(s) for the following tests are stored in
-    tests/test_data/calculate_total_min_mean_max.py
+    tests/test_data/calculate/min_mean_max_data.py
     """
 
     def test_value_error_raised_when_invalid_groupby_columns_provided(self):
@@ -102,14 +103,12 @@ class TestMinMeanMax:
 
     def test_min_mean_max_correct_when_groupd_by_only_gene(self):
         calculated_values = calculate.min_mean_max(
-            coverage_data=calculate_min_mean_max_data.input_calculated_columns_df(),
+            coverage_data=min_mean_max_data.input_calculated_columns_df(),
             group_by_cols=("gene",),
             join=False,
         )
 
-        expected_values = (
-            calculate_min_mean_max_data.expected_grouped_by_gene_df()
-        )
+        expected_values = min_mean_max_data.expected_grouped_by_gene_df()
 
         pl_testing.assert_frame_equal(
             calculated_values, expected_values, check_row_order=False
@@ -117,13 +116,13 @@ class TestMinMeanMax:
 
     def test_min_mean_max_correct_when_grouped_by_gene_and_region(self):
         calculated_values = calculate.min_mean_max(
-            coverage_data=calculate_min_mean_max_data.input_calculated_columns_df(),
+            coverage_data=min_mean_max_data.input_calculated_columns_df(),
             group_by_cols=("gene", "region"),
             join=False,
         )
 
         expected_values = (
-            calculate_min_mean_max_data.expected_grouped_by_gene_and_region_df()
+            min_mean_max_data.expected_grouped_by_gene_and_region_df()
         )
 
         pl_testing.assert_frame_equal(
@@ -134,14 +133,12 @@ class TestMinMeanMax:
 
     def test_columns_correct_when_joined_to_input_dataframe(self):
         calculated_values = calculate.min_mean_max(
-            coverage_data=calculate_min_mean_max_data.input_calculated_columns_df(),
+            coverage_data=min_mean_max_data.input_calculated_columns_df(),
             group_by_cols=("gene", "region"),
             join=True,
         )
 
-        expected_values = (
-            calculate_min_mean_max_data.expected_joined_to_input_df()
-        )
+        expected_values = min_mean_max_data.expected_joined_to_input_df()
 
         pl_testing.assert_frame_equal(
             calculated_values,
