+++ tests/test_data/calculate/min_mean_max_data.py
